changed a few of the files and where the code was located, added thr chat intent and now going to bed, tmr work on other intents and info retrival
