+++ experimently/app/src/main/res/layout/chat_question_content.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/userQuestion"
+        android:layout_width="match_parent"
+        android:layout_height="86dp"
+        android:text="Question" />
+
+    <TextView
+        android:id="@+id/userQuestionID"
+        android:layout_width="match_parent"
+        android:layout_height="33dp"
+        android:text="UserID" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="horizontal">
+
+        <TextView
+            android:id="@+id/userQuestionDate"
+            android:layout_width="291dp"
+            android:layout_height="37dp"
+            android:layout_weight="1"
+            android:text="Date" />
+
+        <TextView
+            android:id="@+id/userQuestionReplies"
+            android:layout_width="wrap_content"
+            android:layout_height="36dp"
+            android:layout_weight="1"
+            android:text="Reply Count" />
+    </LinearLayout>
+</LinearLayout>
